Separate add_sample_data function
